More progress with the webassembly part and started working on the WS server

+++ backend/nbody/src/lib.rs
@@ -2,4 +2,4 @@ pub mod physics;
 pub mod quadtree;
 pub mod simulation;
 
-const SMALL: f32 = 1e-5;
+const SMALL: f64 = 1e-5;
